[UI/searchcondition] - Fix layout and routs
Changes:
- Change searchcondition styles from home.js to searchconditin.js
- Change link to searchcondition styles from home.blade.php to app.blade.php

Additions:
- Add routes to home.blade.php
- Add singin/signup and username nav-item in navbar.blade.php
