+++ routes/web.php
@@ -25,6 +25,10 @@ use App\Http\Controllers\Admin\FollowerFollowingListController;
 
 Auth::routes();
 
+//Home Routes
+Route::get('/', [HomeController::class, 'index'])->name('home');
+Route::get('/home', [HomeController::class, 'index'])->name('home');
+
 
 //Seller Routes
 Route::get('/seller/guide', [SellerController::class, 'guide'])->name('seller.guide');
